Removed Installation verification steps and added reference to Code Samples.
